kvserver: move mma replica related methods to its own files
This commit moves the mma specific *Replica methods to mma_replica_store.go for
better organization. No functional changes have been made yet - only a
copy-paste of the methods. In the future, we plan to introduce type aliases in
mma_replica_store,go to clarify the separation.
