+++ pkg/kv/kvserver/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "merge_queue.go",
         "metric_rules.go",
         "metrics.go",
+        "mma_replica_store.go",
         "mma_store_rebalancer.go",
         "mvcc_gc_queue.go",
         "queue.go",
